Chore/sync fork (#2)
* fix: minor misspellings (#1450)

* fix: minor misspellings

* revert change for `-Bbuild`

* Change LSP Keybindings to Match the Default `gr` Bindings Introduced in Neovim 0.11 (#1427)

* refactor: change LSP keybindings to the default gr bindings introduced in 0.11

* refactor: modify existing LSP functions to follow  convention

* Remove Telescope `0.1` branch lock (#1448)

* feat: switch nvim-cmp for blink.cmp (#1426)

* Change to Mason's new address (#1516)

* feat: switch vim-sleuth for guess-indent.nvim (#1512)

* Replace vim.opt with vim.o (#1495)

* Replace vim.opt with vim.o

Because it offers a nicer interface and info on hover.
For now leave vim.opt when using the table interface (until vim.o
with tables is implemented)

* Add type hint for vim.opt.rtp

* Add a comment about using vim.opt instead of vim.o

* don't lazy-load neo-tree so netrw hijacking on startup works (#1489)

* README: mention fd-find in requirements (#1477)

Fixes #1476

* fix: rename vim.highlight.on_yank to vim.hl.on_yank (#1482)

The functions of vim.highlight were renamed to vim.hl on commit
18b43c331d8a0ed87d7cbefe2a18543b8e4ad360 of neovim, which was applied
with the release of nvim version 0.11.

Now, the use of vim.highlight is deprecated, and instead, one should
use vim.hl functions.
In practice, vim.highlight is still working, however, asking for help
for vim.highlight.on_yank fails (E149), while asking for help for
vim.hl.on_yank works as expected. So, by updating the used function, a
new user will have easier time looking getting the relevant help.

Co-authored-by: Omri Sarig <omri.sarig@prevas.dk>

* Update remaining Mason's old address (#1530)

* Initial plugin setup (#1)

This adds a basic layer of plugins. But more language and workflow specific plugins are to follow possibly.

---------

Co-authored-by: dasvh <dasvh@protonmail.com>
Co-authored-by: Theo P. <63016528+theopn@users.noreply.github.com>
Co-authored-by: Dmytro Onypko <vaporif@gmail.com>
Co-authored-by: Liam Dyer <liamcdyer@gmail.com>
Co-authored-by: guru245 <guru245@users.noreply.github.com>
Co-authored-by: Ori Perry <48057913+oriori1703@users.noreply.github.com>
Co-authored-by: pynappo <lehtien.david@gmail.com>
Co-authored-by: Damjan 9000 <damjan.9000@gmail.com>
Co-authored-by: Omri Sarig <omri.sarig13@gmail.com>
Co-authored-by: Omri Sarig <omri.sarig@prevas.dk>
Co-authored-by: Carlos Calla Alarcón <calla.alarcon@gmail.com>